@@ -10,7 +10,8 @@ from quacc.recipes.emt.core import static_job
 from quacc.wflow_tools.customizers import customize_funcs
 
 if TYPE_CHECKING:
-    from typing import Any, Callable
+    from collections.abc import Callable
+    from typing import Any
 
     from ase.atoms import Atoms
 
@@ -26,6 +27,7 @@ def phonon_flow(
         tuple[tuple[int, int, int], tuple[int, int, int], tuple[int, int, int]] | None
     ) = None,
     displacement: float = 0.01,
+    fixed_atom_indices: list[int] | None = None,
     t_step: float = 10,
     t_min: float = 0,
     t_max: float = 1000,
@@ -62,6 +64,11 @@ def phonon_flow(
         value specified by `min_lengths`.
     displacement
         Atomic displacement (A).
+    fixed_atom_indices
+        Indices of fixed atoms. These atoms will not be displaced
+        during the phonon calculation. Useful for adsorbates on
+        surfaces with weak coupling etc. Important approximation,
+        use with caution.
     t_step
         Temperature step (K).
     t_min
@@ -95,6 +102,7 @@ def phonon_flow(
         symprec=symprec,
         min_lengths=min_lengths,
         supercell_matrix=supercell_matrix,
+        fixed_atom_indices=fixed_atom_indices,
         displacement=displacement,
         t_step=t_step,
         t_min=t_min,
