@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-import logging
 from copy import deepcopy
+from logging import getLogger
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     from quacc.types import AdsSiteFinderKwargs, FindAdsSitesKwargs
 
 
-logger = logging.getLogger(__name__)
+LOGGER = getLogger(__name__)
 
 
 def flip_atoms(
