@@ -2,8 +2,8 @@
 
 from __future__ import annotations
 
-import logging
 from importlib.util import find_spec
+from logging import getLogger
 from typing import TYPE_CHECKING
 
 import numpy as np
@@ -28,7 +28,7 @@ if TYPE_CHECKING:
     if has_atomate2:
         from atomate2.vasp.jobs.base import BaseVaspMaker
 
-logger = logging.getLogger(__name__)
+LOGGER = getLogger(__name__)
 
 
 def get_param_swaps(
@@ -63,12 +63,12 @@ def get_param_swaps(
     if (
         not calc.int_params["lmaxmix"] or calc.int_params["lmaxmix"] < 6
     ) and max_Z > 56:
-        logger.info("Recommending LMAXMIX = 6 because you have f electrons.")
+        LOGGER.info("Recommending LMAXMIX = 6 because you have f electrons.")
         calc.set(lmaxmix=6)
     elif (
         not calc.int_params["lmaxmix"] or calc.int_params["lmaxmix"] < 4
     ) and max_Z > 20:
-        logger.info("Recommending LMAXMIX = 4 because you have d electrons.")
+        LOGGER.info("Recommending LMAXMIX = 4 because you have d electrons.")
         calc.set(lmaxmix=4)
 
     if (
@@ -78,7 +78,7 @@ def get_param_swaps(
         or calc.dict_params["ldau_luj"]
         or calc.string_params["metagga"]
     ) and not calc.bool_params["lasph"]:
-        logger.info(
+        LOGGER.info(
             "Recommending LASPH = True because you have a +U, vdW, meta-GGA, or hybrid calculation."
         )
         calc.set(lasph=True)
@@ -86,14 +86,14 @@ def get_param_swaps(
     if calc.string_params["metagga"] and (
         not calc.string_params["algo"] or calc.string_params["algo"].lower() != "all"
     ):
-        logger.info("Recommending ALGO = All because you have a meta-GGA calculation.")
+        LOGGER.info("Recommending ALGO = All because you have a meta-GGA calculation.")
         calc.set(algo="all")
 
     if calc.bool_params["lhfcalc"] and (
         not calc.string_params["algo"]
         or calc.string_params["algo"].lower() not in ["all", "damped", "normal"]
     ):
-        logger.info("Recommending ALGO = Normal because you have a hybrid calculation.")
+        LOGGER.info("Recommending ALGO = Normal because you have a hybrid calculation.")
         calc.set(algo="normal")
 
     if (
@@ -101,7 +101,7 @@ def get_param_swaps(
         and (calc.int_params["ismear"] and calc.int_params["ismear"] < 0)
         and (calc.int_params["nsw"] and calc.int_params["nsw"] > 0)
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending ISMEAR = 1 and SIGMA = 0.1 because you are likely relaxing a metal."
         )
         calc.set(ismear=1, sigma=0.1)
@@ -114,7 +114,7 @@ def get_param_swaps(
             or (calc.float_params["kspacing"] and calc.float_params["kspacing"] <= 0.5)
         )
     ):
-        logger.info("Recommending ISMEAR = -5 because you have a static calculation.")
+        LOGGER.info("Recommending ISMEAR = -5 because you have a static calculation.")
         calc.set(ismear=-5)
 
     if (
@@ -122,7 +122,7 @@ def get_param_swaps(
         and np.prod(calc.kpts) < 4
         and calc.float_params["kspacing"] is None
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending ISMEAR = 0 because you don't have enough k-points for ISMEAR = -5."
         )
         calc.set(ismear=0)
@@ -132,13 +132,13 @@ def get_param_swaps(
         and calc.float_params["kspacing"] > 0.5
         and calc.int_params["ismear"] == -5
     ):
-        logger.info(
+        LOGGER.info(
             "Recocmmending ISMEAR = 0 because KSPACING is likely too large for ISMEAR = -5."
         )
         calc.set(ismear=0)
 
     if pmg_kpts and pmg_kpts.get("line_density") and calc.int_params["ismear"] != 0:
-        logger.info(
+        LOGGER.info(
             "Recommending ISMEAR = 0 and SIGMA = 0.01 because you are doing a line mode calculation."
         )
         calc.set(ismear=0, sigma=0.01)
@@ -146,7 +146,7 @@ def get_param_swaps(
     if calc.int_params["ismear"] == 0 and (
         not calc.float_params["sigma"] or calc.float_params["sigma"] > 0.05
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending SIGMA = 0.05 because ISMEAR = 0 was requested with SIGMA > 0.05."
         )
         calc.set(sigma=0.05)
@@ -156,7 +156,7 @@ def get_param_swaps(
         and calc.int_params["nsw"] > 0
         and calc.bool_params["laechg"]
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending LAECHG = False because you have NSW > 0. LAECHG is not compatible with NSW > 0."
         )
         calc.set(laechg=False)
@@ -164,11 +164,11 @@ def get_param_swaps(
     if calc.int_params["ldauprint"] in (None, 0) and (
         calc.bool_params["ldau"] or calc.dict_params["ldau_luj"]
     ):
-        logger.info("Recommending LDAUPRINT = 1 because LDAU = True.")
+        LOGGER.info("Recommending LDAUPRINT = 1 because LDAU = True.")
         calc.set(ldauprint=1)
 
     if calc.special_params["lreal"] and len(input_atoms) < 30:
-        logger.info(
+        LOGGER.info(
             "Recommending LREAL = False because you have a small system (< 30 atoms/cell)."
         )
         calc.set(lreal=False)
@@ -177,7 +177,7 @@ def get_param_swaps(
         calc.int_params["ispin"] == 2
         or np.any(input_atoms.get_initial_magnetic_moments() != 0)
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending LORBIT = 11 because you have a spin-polarized calculation."
         )
         calc.set(lorbit=11)
@@ -191,7 +191,7 @@ def get_param_swaps(
         or calc.bool_params["lepsilon"] is True
         or calc.int_params["ibrion"] in [5, 6, 7, 8]
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending NCORE = 1 because NCORE/NPAR is not compatible with this job type."
         )
         calc.set(ncore=1, npar=None)
@@ -201,19 +201,19 @@ def get_param_swaps(
         and calc.int_params["kpar"] > np.prod(calc.kpts)
         and calc.float_params["kspacing"] is None
     ):
-        logger.info(
+        LOGGER.info(
             "Recommending KPAR = 1 because you have too few k-points to parallelize."
         )
         calc.set(kpar=1)
 
     if calc.bool_params["lhfcalc"] is True and calc.int_params["isym"] in (1, 2):
-        logger.info(
+        LOGGER.info(
             "Recommending ISYM = 3 because you are running a hybrid calculation."
         )
         calc.set(isym=3)
 
     if calc.bool_params["lsorbit"]:
-        logger.info(
+        LOGGER.info(
             "Recommending ISYM = -1 because you are running an SOC calculation."
         )
         calc.set(isym=-1)
@@ -222,7 +222,7 @@ def get_param_swaps(
         (calc.int_params["ncore"] and calc.int_params["ncore"] > 1)
         or (calc.int_params["npar"] and calc.int_params["npar"] > 1)
     ) and (calc.bool_params["lelf"] is True):
-        logger.info(
+        LOGGER.info(
             "Recommending NPAR = 1 because NCORE/NPAR is not compatible with this job type."
         )
         calc.set(npar=1, ncore=None)
@@ -239,7 +239,7 @@ def get_param_swaps(
     if changed_parameters := {
         k: new_parameters[k] for k in set(new_parameters) - set(user_calc_params)
     }:
-        logger.info(
+        LOGGER.info(
             f"The following parameters were changed: {sort_dict(changed_parameters)}"
         )
 
@@ -449,7 +449,11 @@ class MPtoASEConverter:
         )
         self.incar_dict = input_set.incar
         self.pmg_kpts = input_set.kpoints
-        self.potcar_symbols = input_set.potcar
+        self.potcar_symbols = (
+            input_set.potcar.split("\n")
+            if isinstance(input_set.potcar, str)
+            else input_set.potcar
+        )
         self.potcar_functional = input_set_generator.potcar_functional
         self.poscar = input_set.poscar
         return self._convert()
