@@ -16,12 +16,12 @@ has_obabel = bool(find_spec("openbabel"))
 FILE_DIR = Path(__file__).parent
 
 
-@pytest.fixture()
+@pytest.fixture
 def test_atoms():
     return read(FILE_DIR / "test.xyz")
 
 
-@pytest.fixture()
+@pytest.fixture
 def os_atoms():
     return read(FILE_DIR / "OS_test.xyz")
 
