@@ -7,7 +7,7 @@ import pytest
 pytest.importorskip("phonopy")
 pytest.importorskip("seekpath")
 
-from ase.build import bulk
+from ase.build import bulk, molecule
 
 from quacc.recipes.emt.phonons import phonon_flow
 
@@ -58,3 +58,33 @@ def test_phonon_flow_v3(tmp_path, monkeypatch):
     assert output["results"]["force_constants"].shape == (8, 8, 3, 3)
     assert "mesh_properties" in output["results"]
     assert output["atoms"] == atoms
+
+
+def test_phonon_flow_fixed(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+    atoms = molecule("H2", vacuum=20.0)
+    output = phonon_flow(atoms, min_lengths=5.0)
+
+    atoms1 = molecule("H2", vacuum=20.0)
+    atoms2 = molecule("H2", vacuum=20.0)
+
+    atoms2.positions += [10, 10, 10]
+
+    atoms = atoms1 + atoms2
+
+    output_fixed = phonon_flow(atoms, fixed_atom_indices=[2, 3], min_lengths=1.0)
+    # Should be very close but not exactly the same, also check the size is correct
+    assert output["results"]["mesh_properties"]["frequencies"] == pytest.approx(
+        output_fixed["results"]["mesh_properties"]["frequencies"], rel=0.0, abs=1e-5
+    )
+
+    assert len(output_fixed["displaced_atoms"]) == 2
+    assert len(output_fixed["non_displaced_atoms"]) == 2
+
+    output_fixed_wrong = phonon_flow(atoms, fixed_atom_indices=[0, 2], min_lengths=1.0)
+
+    assert output_fixed_wrong["results"]["mesh_properties"][
+        "frequencies"
+    ] != pytest.approx(
+        output_fixed["results"]["mesh_properties"]["frequencies"], rel=0.0, abs=1e-5
+    )
