@@ -699,7 +699,6 @@ def test_mp_gga_relax_job(patch_nonmetallic_taskdoc):
         "ispin": 2,
         "kpts": (5, 11, 11),
         "lasph": True,
-        "lcharg": False,
         "ldau": True,
         "ldauj": [0, 0],
         "ldaul": [0, 2],
@@ -735,7 +734,7 @@ def test_mp_gga_static_job(patch_nonmetallic_taskdoc):
         "gamma": True,
         "ismear": -5,
         "ispin": 2,
-        "kpts": (5, 11, 11),
+        "kpts": (6, 13, 13),
         "lasph": True,
         "lcharg": True,
         "ldau": True,
@@ -778,7 +777,6 @@ def test_mp_gga_relax_flow(tmp_path, patch_nonmetallic_taskdoc):
             "ispin": 2,
             "kpts": (5, 11, 11),
             "lasph": True,
-            "lcharg": False,
             "ldau": True,
             "ldauj": [0, 0],
             "ldaul": [0, 2],
@@ -809,7 +807,7 @@ def test_mp_gga_relax_flow(tmp_path, patch_nonmetallic_taskdoc):
             "gamma": True,
             "ismear": -5,
             "ispin": 2,
-            "kpts": (5, 11, 11),
+            "kpts": (6, 13, 13),
             "lasph": True,
             "lcharg": True,
             "ldau": True,
