@@ -10,7 +10,7 @@ from pymatgen.io.ase import AseAtomsAdaptor
 from quacc.schemas.atoms import atoms_to_metadata
 
 
-@pytest.fixture()
+@pytest.fixture
 def test_cifs():
     file_dir = Path(__file__).parent
 
