@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-import logging
 import os
+from logging import WARNING, getLogger
 from pathlib import Path
 from shutil import copytree, move
 
@@ -13,18 +13,18 @@ from monty.json import MontyDecoder, jsanitize
 from quacc.calculators.vasp import Vasp
 from quacc.schemas.vasp import VaspSummarize
 
-LOGGER = logging.getLogger(__name__)
+LOGGER = getLogger(__name__)
 LOGGER.propagate = True
 
 
-@pytest.fixture()
+@pytest.fixture
 def run1():
     FILE_DIR = Path(__file__).parent
 
     return FILE_DIR / "test_files" / "vasp_run1"
 
 
-@pytest.fixture()
+@pytest.fixture
 def mp_run1():
     FILE_DIR = Path(__file__).parent
 
@@ -217,7 +217,7 @@ def test_summarize_mp_bad(monkeypatch, run1, tmp_path, caplog):
     p = tmp_path / "vasp_run"
     copytree(run1, p)
     atoms = read(p / "OUTCAR.gz")
-    with caplog.at_level(logging.WARNING):
+    with caplog.at_level(WARNING):
         VaspSummarize(directory=p, report_mp_corrections=True).run(atoms)
     assert "invalid run type" in caplog.text
 
@@ -227,7 +227,7 @@ def test_no_bader(tmp_path, monkeypatch, run1, caplog):
     p = tmp_path / "vasp_run"
     copytree(run1, p)
     atoms = read(p / "OUTCAR.gz")
-    with caplog.at_level(logging.WARNING):
+    with caplog.at_level(WARNING):
         VaspSummarize(directory=p, run_bader=True, run_chargemol=False).run(atoms)
     assert "Bader analysis could not be performed." in caplog.text
 
@@ -237,6 +237,6 @@ def test_no_chargemol(tmp_path, monkeypatch, run1, caplog):
     p = tmp_path / "vasp_run"
     copytree(run1, p)
     atoms = read(p / "OUTCAR.gz")
-    with caplog.at_level(logging.WARNING):
+    with caplog.at_level(WARNING):
         VaspSummarize(directory=p, run_bader=False, run_chargemol=True).run(atoms)
     assert "Chargemol analysis could not be performed." in caplog.text
