@@ -1,13 +1,13 @@
 from __future__ import annotations
 
-import logging
+from logging import WARNING, getLogger
 
 import pytest
 
 from quacc import Remove
 from quacc.utils.dicts import finalize_dict, recursive_dict_merge, remove_dict_entries
 
-LOGGER = logging.getLogger(__name__)
+LOGGER = getLogger(__name__)
 LOGGER.propagate = True
 
 
@@ -73,7 +73,7 @@ def test_recursive_dict_merge2():
 def test_recursive_dict_merge_verbose(caplog):
     defaults = {"a": 1, "b": {"a": 1, "b": 2}}
     calc_swaps = {"a": Remove, "b": {"b": 3, "d": 1}}
-    with caplog.at_level(logging.WARNING):
+    with caplog.at_level(WARNING):
         recursive_dict_merge(defaults, calc_swaps, verbose=True)
         assert "Overwriting key 'a'" in caplog.text
         assert "Overwriting key 'b' to: '3'" in caplog.text
