@@ -1,9 +1,9 @@
 from __future__ import annotations
 
 import gzip
-import logging
 import os
 import time
+from logging import WARNING, getLogger
 from pathlib import Path
 
 import pytest
@@ -15,7 +15,7 @@ from quacc.utils.files import (
     make_unique_dir,
 )
 
-LOGGER = logging.getLogger(__name__)
+LOGGER = getLogger(__name__)
 LOGGER.propagate = True
 
 
@@ -119,7 +119,7 @@ def test_copy_decompress_files_v4(tmp_path):
 
 
 def test_copy_decompress_files_from_dir_warning(caplog):
-    with caplog.at_level(logging.WARNING):
+    with caplog.at_level(WARNING):
         copy_decompress_files("fake", "file", "test")
     assert "Cannot find file" in caplog.text
 
