@@ -27,7 +27,7 @@ def test_patch():
         future1 = task1(val)
         return task2(future1["result"])
 
-    assert workflow(1).result() == {"input": 100, "result": 20000}
+    assert workflow(1) == {"input": 100, "result": 20000}
 
 
 def test_patch_local():
@@ -60,7 +60,7 @@ def test_prefect_decorators1(tmp_path, monkeypatch):
     def add_flow(a, b):
         return add(a, b)
 
-    assert add_flow(1, 2).result() == 3
+    assert add_flow(1, 2) == 3
 
 
 def test_prefect_decorators2(tmp_path, monkeypatch):
@@ -78,7 +78,7 @@ def test_prefect_decorators2(tmp_path, monkeypatch):
     def workflow(a, b, c):
         return mult(add(a, b), c)
 
-    assert workflow(1, 2, 3).result() == 9
+    assert workflow(1, 2, 3) == 9
 
 
 def test_prefect_decorators3(tmp_path, monkeypatch):
@@ -102,7 +102,7 @@ def test_prefect_decorators3(tmp_path, monkeypatch):
         result2 = make_more(result1)
         return add_distributed(result2, c)
 
-    assert [r.result() for r in dynamic_workflow(1, 2, 3)] == [6, 6, 6]
+    assert dynamic_workflow(1, 2, 3) == [6, 6, 6]
 
 
 def test_prefect_decorators4(tmp_path, monkeypatch):
@@ -126,7 +126,7 @@ def test_prefect_decorators4(tmp_path, monkeypatch):
         result2 = make_more(result1)
         return add_distributed2(result2, c, add)
 
-    assert [r.result() for r in dynamic_workflow2(1, 2, 3)] == [6, 6, 6]
+    assert dynamic_workflow2(1, 2, 3) == [6, 6, 6]
 
 
 def test_prefect_decorators5(tmp_path, monkeypatch):
@@ -152,7 +152,7 @@ def test_prefect_decorators5(tmp_path, monkeypatch):
         result4 = add_distributed(result3, c)
         return add(result4[0], c)
 
-    assert dynamic_workflow3(1, 2, 3).result() == 12
+    assert dynamic_workflow3(1, 2, 3) == 12
 
 
 def test_prefect_decorators_local(tmp_path, monkeypatch):
@@ -239,10 +239,6 @@ def test_state_patch():
 
         return 2
 
-    # Note: No need for `.result()` here because we are returning data, not a PrefectFuture
-    # Doing `my_flow(return_state=True)` would force the need for a `.result()`. In general,
-    # Prefect flows return the object as-is unless `return_state=True` is set. It's just
-    # that in general throughout quacc, we are often returning an unresolved `PrefectFuture`
     assert my_flow() == 2
 
 
@@ -264,3 +260,41 @@ def test_concurrency():
     t1 = time.time()
     dt = t1 - t0
     assert dt < 10
+
+
+async def test_prefect_decorators_async(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+
+    @job
+    async def add(a, b):
+        return a + b
+
+    @flow
+    async def add_flow(a, b):
+        return add(a, b)
+
+    assert (await add_flow(1, 2)) == 3
+
+
+async def test_prefect_decorators3_async(tmp_path, monkeypatch):
+    monkeypatch.chdir(tmp_path)
+
+    @job
+    async def add(a, b):
+        return a + b
+
+    @job
+    async def make_more(val):
+        return [val] * 3
+
+    @subflow
+    async def add_distributed(vals, c):
+        return [add(val, c) for val in vals]
+
+    @flow
+    async def dynamic_workflow(a, b, c):
+        result1 = add(a, b)
+        result2 = make_more(result1)
+        return await add_distributed(result2, c)
+
+    assert (await dynamic_workflow(1, 2, 3)) == [6, 6, 6]
