@@ -10,7 +10,7 @@ from quacc.recipes.emt.core import relax_job, static_job  # skipcq: PYL-C0412
 from quacc.recipes.emt.slabs import bulk_to_slabs_flow  # skipcq: PYL-C0412
 
 
-@pytest.fixture()
+@pytest.fixture
 def scheduler():
     return redun.Scheduler()
 
